+++ src/Request/Model/AttachTeamToGroup.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Easir\SDK\Request\Model;
+
+use Easir\SDK\Request\Model;
+
+class AttachTeamToGroup extends Model
+{
+    /**
+     * @var int
+     */
+    public $group_id;
+    /**
+     * @var int
+     */
+    public $team_id;
+}