+++ src/Request/Model/AttachUserToTeam.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Easir\SDK\Request\Model;
+
+use Easir\SDK\Request\Model;
+
+class AttachUserToTeam extends Model
+{
+    /**
+     * @var int
+     */
+    public $team_id;
+    /**
+     * @var int
+     */
+    public $user_id;
+}