+++ src/Request/Model/CreateGroup.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Easir\SDK\Request\Model;
+
+use Easir\SDK\Request\Model;
+
+class CreateGroup extends Model
+{
+    /**
+     * @var string
+     */
+    public $name;
+}